+++ jstarcraft-ai-model/src/main/java/com/jstarcraft/ai/model/neuralnetwork/condition/EpsilonCondition.java
@@ -19,27 +19,27 @@ import com.jstarcraft.ai.math.structure.matrix.MathMatrix;
  */
 public class EpsilonCondition implements Condition {
 
-    private double epsilon = MathUtility.EPSILON;
+    private float epsilon = MathUtility.EPSILON;
 
-    private double tolerance = MathUtility.EPSILON;
+    private float tolerance = MathUtility.EPSILON;
 
     public EpsilonCondition() {
     }
 
-    public EpsilonCondition(double epsilon, double tolerance) {
+    public EpsilonCondition(float epsilon, float tolerance) {
         this.epsilon = epsilon;
         this.tolerance = tolerance;
     }
 
     @Override
-    public boolean stop(double newScore, double oldScore, Map<String, MathMatrix> gradients) {
+    public boolean stop(float newScore, float oldScore, Map<String, MathMatrix> gradients) {
         // special case for initial termination, ignore
-        if (newScore == 0D && oldScore == 0D) {
+        if (newScore == 0F && oldScore == 0F) {
             return false;
         }
 
-        double oldAbsolute = FastMath.abs(oldScore);
-        double newAbsolute = FastMath.abs(newScore);
+        float oldAbsolute = FastMath.abs(oldScore);
+        float newAbsolute = FastMath.abs(newScore);
 
         return 2D * FastMath.abs(oldScore - newScore) <= tolerance * (oldAbsolute + newAbsolute + epsilon);
     }
